Grad-CAM: Visualizing Neural Network Decisions
In this project, I explore the fascinating world of convolutional neural networks (CNNs) and their interpretability through a technique called Grad-CAM (Gradient-weighted Class Activation Mapping). This method allows us to visualize and understand which parts of an image are driving the network’s decisions, offering a glimpse into the "black box" nature of deep learning models.

Grad-CAM works by calculating the gradient of the target class with respect to the final convolutional layer, and then using this information to generate a heatmap. This heatmap highlights the areas of the input image that are most important for the model's prediction.

By implementing Grad-CAM, I provide insights into:

The inner workings of CNNs.

Which features are learned by the network and how they contribute to the final classification.

For the data set i took an image of an Elephant and tried to create a heatmap over it.

Feel free to explore the notebook and see how Grad-CAM can make deep learning more interpretable and accessible!